remove unneccessary deep clone
Deep cloning the output object using `JSON.stringify` & `JSON.parse` eliminates the ability to use complex data types. Removing this deep clone entirely only fails one test, changing a behavior in which a `undefined` array value gets converted to `null` in the process of deep cloning with JSON. Since this behavior appears to be a side effect and not the primary target of the test, I've adjusted the test slightly to stay true to the description of the test but not rely on this side effect behavior of converting to and from JSON in validating the outcome.