Implement custom functions instead of user set
This Object Mapper will be used diferently from the fork origin. We will
not be accepting the transform funciton from the user. Rather we will
have a set list of transforms available to the user. Our implementation
of the Object mapper is completely configurable via json files therefore
to accept transformation functions from the user would require running
user defined functions directly which is not advised as a malicious user
could do damage on the server our application is running on.

This method will allow catered for transformations without risk to the
server. If users require more functions than currently implemented then
they can open a PR to this repository or fork it and the
openhim-mediator-mapping

OHM-870